refactor(test-suite): keep rollout cleanup narrow
Limit the rollout boundary cleanup to explicit previous-contract snapshots and upgrade planning extraction.
